useMatchManagementフックに勝者設定機能を追加し、Mainコンポーネントでの試合結果管理を改善
